add "stop generate" button

+++ components/ChatHistory.jsx
@@ -5,6 +5,10 @@ const IconEdit = ({ className, onClick }) => <svg onClick={onClick} stroke="curr
 
 const IconDelete = ({ className, onClick }) => <svg onClick={onClick} stroke="currentColor" fill="none" strokeWidth="2" viewBox="0 0 24 24" strokeLinecap="round" strokeLinejoin="round" className={"h-4 w-4 " + className} height="1em" width="1em" xmlns="http://www.w3.org/2000/svg"><polyline points="3 6 5 6 21 6"></polyline><path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path><line x1="10" y1="11" x2="10" y2="17"></line><line x1="14" y1="11" x2="14" y2="17"></line></svg>
 
+const IconUser = ({ className, onClick }) => <svg onClick={onClick} stroke="currentColor" fill="none" strokeWidth="2" viewBox="0 0 24 24" strokeLinecap="round" strokeLinejoin="round" className={"h-4 w-4 " + className} height="1em" width="1em" xmlns="http://www.w3.org/2000/svg"><path d="M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2"></path><circle cx="12" cy="7" r="4"></circle></svg>
+
+const IconOpen = ({ className, onClick }) => <svg onClick={onClick} stroke="currentColor" fill="none" stroke-width="2" viewBox="0 0 24 24" strokeLinecap="round" strokeLinejoin="round" className={"h-4 w-4 " + className} height="1em" width="1em" xmlns="http://www.w3.org/2000/svg"><path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path><polyline points="15 3 21 3 21 9"></polyline><line x1="10" y1="14" x2="21" y2="3"></line></svg>
+
 function ChatHistory({
   chats,
   userText,
@@ -136,33 +140,28 @@ function ChatHistory({
                 ))}
               </div>
             </div>
-            {chats.length > 0 && (
+            {chats.length > 0 && <>
               <button
                 className="flex py-3 px-3 items-center gap-3 rounded-md hover:bg-gray-500/10 transition-colors duration-200 text-white cursor-pointer text-sm"
                 onClick={() => {
-                  deleteChats();
+                  if (window.confirm('This will delete all conversations. Are you sure?')) {
+                    deleteChats();
+                  }
                 }}
               >
-                <svg
-                  stroke="currentColor"
-                  fill="none"
-                  strokeWidth="2"
-                  viewBox="0 0 24 24"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  className="h-4 w-4"
-                  height="1em"
-                  width="1em"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <polyline points="3 6 5 6 21 6"></polyline>
-                  <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2"></path>
-                  <line x1="10" y1="11" x2="10" y2="17"></line>
-                  <line x1="14" y1="11" x2="14" y2="17"></line>
-                </svg>
+                <IconDelete />
                 Clear conversations
               </button>
-            )}
+              {/* <button
+                className="flex py-3 px-3 items-center gap-3 rounded-md hover:bg-gray-500/10 transition-colors duration-200 text-white cursor-pointer text-sm"
+                onClick={() => {
+                  // TODO
+                }}
+              >
+                <IconUser />
+                Settings
+              </button> */}
+            </>}
           </nav>
         </div>
       </div>
